feat!: terminate with original signal unless `useExit0` option passed
BREAKING CHANGE: Process will now exit with the signal it received, instead of 0/1.

I believe it is more expected for a process to exit with the signal it was passed, rather than exiting with a generic 0 or 1. There are cases where a 0 exit code is wanted, so I added an option to allow for this behavior to continue. Note that I have inverted the default from what it previously was, thus this should be considered a breaking change.